Refactor into core, CLI, and API modules
