+++ langgraph.json
@@ -1,6 +1,1 @@
-{
-  "dependencies": ["."],
-  "graphs": {
-    "mcp_greeter": "./mcp_graph_greeter.py:graph_factory"
-  }
-}
+{"dependencies": ["."], "graphs": {"mcp_greeter": "./cli.py:graph_factory"}}
